NOW WE ARE TALKING: added single cuotes and double cutes correctly(as far as my knowledge goes)

+++ minishell/libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: kmb <kmb@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/11 22:20:09 by akambou           #+#    #+#             */
-/*   Updated: 2024/01/06 16:00:24 by kmb              ###   ########.fr       */
+/*   Updated: 2024/01/20 21:49:12 by kmb              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len);
 void	*ft_memchr(const void *s, int c, size_t n);
 void	*ft_memset(void *b, int c, size_t len);
 /*-------------------------------STRINGS----------------------------------*/
-size_t	ft_strlen(const char *str);
+int	ft_strlen(const char *str);
 size_t	ft_strlcpy(char *dest, const char *src, size_t len);
 size_t	ft_strlcat(char *dest, const char *src, size_t size);
 size_t  ft_strspn(char *str, const char *str2);
